requirements/base.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0

# Database
dj-database-url>=2.1,<3.0
psycopg[binary]>=3.1,<4.0

# Celery & Redis
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# Notifications
pywebpush>=1.14,<2.0
requests>=2.31,<3.0

# Production
whitenoise>=6.6,<7.0
gunicorn>=21.0,<22.0

# Utils
python-dotenv>=1.0,<2.0
